chore(base model): abstract methods to helper class
[#120829713]
